+++ src/hotspot/share/cds/heapShared.hpp
@@ -291,7 +291,7 @@ private:
 
   static GrowableArrayCHeap<oop, mtClassShared>* _pending_roots;
   static GrowableArrayCHeap<OopHandle, mtClassShared>* _root_segments;
-  static size_t _root_segment_max_size;
+  static int _root_segment_max_size;
   static OopHandle _scratch_basic_type_mirrors[T_VOID+1];
   static MetaspaceObjToOopHandleTable* _scratch_java_mirror_table;
   static MetaspaceObjToOopHandleTable* _scratch_references_table;
@@ -424,7 +424,7 @@ private:
   static void init_for_dumping(TRAPS) NOT_CDS_JAVA_HEAP_RETURN;
   static void write_subgraph_info_table() NOT_CDS_JAVA_HEAP_RETURN;
   static void add_root_segment(oop segment_oop) NOT_CDS_JAVA_HEAP_RETURN;
-  static void init_root_segment_max_size(size_t size) NOT_CDS_JAVA_HEAP_RETURN;
+  static void init_root_segment_max_size(int size) NOT_CDS_JAVA_HEAP_RETURN;
   static void serialize_tables(SerializeClosure* soc) NOT_CDS_JAVA_HEAP_RETURN;
 
 #ifndef PRODUCT
